hit api for update data
