+++ src/api/constants.js
@@ -8,4 +8,6 @@ export const METHODS = {
 export const USER_URL = '/user/register';
 export const USER_LOGIN_URL = '/user/login';
 export const TEST_DATA = '/tutorial/addLesson';
+export const UPDATE_DATA = '/user/userUpdate';
+
