+++ src/containers/UserProfile/EditIntro.js
@@ -1,5 +1,9 @@
-import React, { Component } from "react";
+import React, { Component, Fragment } from "react";
 import SimpleReactValidator from "simple-react-validator";
+import ReactNotification from "react-notifications-component";
+import "react-notifications-component/dist/theme.css";
+import addNotification from '../../utilities/index'
+import API from '../../api/index'
 import "./EditIntro.scss";
 
 class EditInto extends Component {
@@ -12,6 +16,9 @@ class EditInto extends Component {
       email: ""
     };
     this.handleUserInput = this.handleUserInput.bind(this);
+    this.notificationDOMRef = React.createRef();
+    this.submitForm = this.submitForm.bind(this);
+    this.formRef = null;
   }
 
   handleUserInput(e) {
@@ -20,65 +27,116 @@ class EditInto extends Component {
     this.setState({ [name]: value });
   }
 
+  submitForm(e) {
+    e.preventDefault();
+    if (this.validator.allValid()) {
+      var { name, email } = this.state;
+      const data = {
+        name,
+        email
+      };
+      API.updateData(data, result => {
+        if (result.status === "200") {
+          //Form reset
+          this.formRef.reset();
+        } else if (
+          result.status === "404" ||
+          result.status === "403" ||
+          result.status === "500"
+        ) {
+          addNotification(
+            this.notificationDOMRef,
+            "Error",
+            "danger",
+            result.message
+          );
+        } else {
+          addNotification(
+            this.notificationDOMRef,
+            "Error",
+            "warning",
+            "Somgthing went wrong"
+          );
+        }
+      }).catch = error => {
+        addNotification(this.notificationDOMRef, "Error", "warning", error);
+      };
+    } else {
+      this.validator.showMessages();
+      // rerender to show messages for the first time
+      this.forceUpdate();
+    }
+  }
+
   render() {
     return (
-      <div className="container">
-        <div className="header">
-          <div className="heading">
-            <h5>Edit Intro</h5>
-          </div>
+      <Fragment>
+        <div>
+          <ReactNotification ref={this.notificationDOMRef} />
         </div>
-        <form
-          className="form"
-          onSubmit={this.submitForm}
-          ref={ref => (this.formRef = ref)}
-        >
-          <div className="feilds" />
-          <div className="name">
-            <label className="labels vertical-spacing">
-              Name <span className="edit-required-indicator">*</span>
-            </label>
-            <br />
-            <input
-            //  value={this.props.name}
-              className="edit-field"
-              name="name"
-              onChange={this.handleUserInput}
-            />
-            <div className="edit-error-msg">
-              {this.validator.message("name", this.state.name, "required")}
+        <div className="container">
+          <div className="header">
+            <div className="heading">
+              <h5>Edit Intro</h5>
             </div>
           </div>
-          <div className="email-div">
-            <label className="labels vertical-spacing">
-              Email <span className="edit-required-indicator">*</span>
-            </label>
-            <br />
-            <input
-            // value={this.props.email}
-              className="edit-field"
-              name="email"
-              type="email"
-              onChange={this.handleUserInput}
-            />
-            <div className="edit-error-msg">
-              {this.validator.message("email", this.state.email, "required")}
+          <form
+            className="form"
+            onSubmit={this.submitForm}
+            ref={ref => (this.formRef = ref)}
+          >
+            <div className="feilds" />
+            <div className="name">
+              <label className="labels vertical-spacing">
+                Name <span className="edit-required-indicator">*</span>
+              </label>
+              <br />
+              <input
+                //  value={this.props.name}
+                className="edit-field"
+                name="name"
+                onChange={this.handleUserInput}
+              />
+              <div className="edit-error-msg">
+                {this.validator.message("name", this.state.name, "required")}
+              </div>
             </div>
-          </div>
-          <div className="btns-div">
-            <button
-              className="btn btn-outline-success"
-              name="cancelBtn"
-              type="submit"
-            >
-              Cancel
-            </button>
-            <button className="btn btn-success" name="saveBtn" type="submit">
-              Save
-            </button>
-          </div>
-        </form>
-      </div>
+            <div className="email-div">
+              <label className="labels vertical-spacing">
+                Email <span className="edit-required-indicator">*</span>
+              </label>
+              <br />
+              <input
+                // value={this.props.email}
+                className="edit-field"
+                name="email"
+                type="email"
+                onChange={this.handleUserInput}
+              />
+              <div className="edit-error-msg">
+                {this.validator.message("email", this.state.email, "required")}
+              </div>
+            </div>
+            <div className="btns-div">
+              <button
+                className="btn btn-outline-success"
+                name="cancelBtn"
+                type="submit"
+              >
+                Cancel
+              </button>
+              <button
+                className="btn btn-success"
+                name="saveBtn"
+                type="submit"
+                onclick={this.saveData}
+              >
+                Save
+              </button>
+            </div>
+          </form>
+        </div>
+      </Fragment>
     );
   }
 }
